test(inventory/aws_ec2): reorganise and expand unit test coverage
Move hostname and availability zone tests to dedicated test files for
better organisation. Add new test coverage for Route53 hostname
matching and Jinja2 filter template formatting.
