+++ tests/unit/plugins/inventory/test_aws_ec2.py
@@ -118,58 +118,6 @@ def test_get_boto_attr_chain(m_compile_values, filter_name, expected):
     assert _get_boto_attr_chain(filter_name, instance) == expected
 
 
-@pytest.mark.parametrize(
-    "hostnames,expected",
-    [
-        ([], "test-instance.ansible.com"),
-        (["private-dns-name"], "test-instance.localhost"),
-        (["tag:os_version"], "RHEL"),
-        (["tag:os_version", "dns-name"], "RHEL"),
-        ([{"name": "Name", "prefix": "Phase"}], "dev_test-instance-01"),
-        ([{"name": "Name", "prefix": "Phase", "separator": "-"}], "dev-test-instance-01"),
-        ([{"name": "Name", "prefix": "OSVersion", "separator": "-"}], "test-instance-01"),
-        ([{"name": "Name", "separator": "-"}], "test-instance-01"),
-        ([{"name": "Name", "prefix": "Phase"}, "private-dns-name"], "dev_test-instance-01"),
-        ([{"name": "Name", "prefix": "Phase"}, "tag:os_version"], "dev_test-instance-01"),
-        (["private-dns-name", "dns-name"], "test-instance.localhost"),
-        (["private-dns-name", {"name": "Name", "separator": "-"}], "test-instance.localhost"),
-        (["private-dns-name", "tag:os_version"], "test-instance.localhost"),
-        (["OSRelease"], None),
-    ],
-)
-@patch("ansible_collections.amazon.aws.plugins.inventory.aws_ec2._get_tag_hostname")
-@patch("ansible_collections.amazon.aws.plugins.inventory.aws_ec2._get_boto_attr_chain")
-def test_inventory_get_preferred_hostname(m_get_boto_attr_chain, m_get_tag_hostname, inventory, hostnames, expected):
-    instance = {
-        "Name": "test-instance-01",
-        "Phase": "dev",
-        "tag:os_version": ["RHEL", "CoreOS"],
-        "another_key": "another_value",
-        "dns-name": "test-instance.ansible.com",
-        "private-dns-name": "test-instance.localhost",
-    }
-
-    inventory._sanitize_hostname = MagicMock()
-    inventory._sanitize_hostname.side_effect = lambda x: x
-
-    m_get_boto_attr_chain.side_effect = lambda pref, instance: instance.get(pref)
-    m_get_tag_hostname.side_effect = lambda pref, instance: instance.get(pref)
-
-    assert expected == inventory._get_preferred_hostname(instance, hostnames)
-
-
-def test_inventory_get_preferred_hostname_failure(inventory):
-    instance = {}
-    hostnames = [{"value": "saome_value"}]
-
-    inventory._sanitize_hostname = MagicMock()
-    inventory._sanitize_hostname.side_effect = lambda x: x
-
-    with pytest.raises(AnsibleError) as err:
-        inventory._get_preferred_hostname(instance, hostnames)
-    assert "A 'name' key must be defined in a hostnames dictionary." in str(err.value)
-
-
 @pytest.mark.parametrize("base_verify_file_return", [True, False])
 @pytest.mark.parametrize(
     "filename,result",
@@ -247,16 +195,6 @@ def test_inventory_build_include_filters(inventory, _options, expected):
     inventory.get_option = MagicMock()
     inventory.get_option.side_effect = _options.get
     assert inventory.build_include_filters() == expected
-
-
-@pytest.mark.parametrize("hostname,expected", [(1, "1"), ("a:b", "a_b"), ("a:/b", "a__b"), ("example", "example")])
-def test_sanitize_hostname(inventory, hostname, expected):
-    assert inventory._sanitize_hostname(hostname) == expected
-
-
-def test_sanitize_hostname_legacy(inventory):
-    inventory._sanitize_group_name = inventory._legacy_script_compatible_group_sanitization
-    assert inventory._sanitize_hostname("a:/b") == "a__b"
 
 
 @pytest.mark.parametrize(
@@ -607,56 +545,6 @@ def test_inventory_get_instances_by_region_failures(m_describe_ec2_instances, in
             inventory._get_instances_by_region(regions, [], strict)
 
 
-@pytest.mark.parametrize(
-    "hostnames,expected",
-    [
-        ([], ["test-instance.ansible.com", "test-instance.localhost"]),
-        (["private-dns-name"], ["test-instance.localhost"]),
-        (["tag:os_version"], ["RHEL", "CoreOS"]),
-        (["tag:os_version", "dns-name"], ["RHEL", "CoreOS", "test-instance.ansible.com"]),
-        ([{"name": "Name", "prefix": "Phase"}], ["dev_test-instance-01"]),
-        ([{"name": "Name", "prefix": "Phase", "separator": "-"}], ["dev-test-instance-01"]),
-        ([{"name": "Name", "prefix": "OSVersion", "separator": "-"}], ["test-instance-01"]),
-        ([{"name": "Name", "separator": "-"}], ["test-instance-01"]),
-        (
-            [{"name": "Name", "prefix": "Phase"}, "private-dns-name"],
-            ["dev_test-instance-01", "test-instance.localhost"],
-        ),
-        ([{"name": "Name", "prefix": "Phase"}, "tag:os_version"], ["dev_test-instance-01", "RHEL", "CoreOS"]),
-        (["private-dns-name", {"name": "Name", "separator": "-"}], ["test-instance.localhost", "test-instance-01"]),
-        (["OSRelease"], []),
-    ],
-)
-@patch("ansible_collections.amazon.aws.plugins.inventory.aws_ec2._get_tag_hostname")
-@patch("ansible_collections.amazon.aws.plugins.inventory.aws_ec2._get_boto_attr_chain")
-def test_inventory_get_all_hostnames(m_get_boto_attr_chain, m_get_tag_hostname, inventory, hostnames, expected):
-    instance = {
-        "Name": "test-instance-01",
-        "Phase": "dev",
-        "tag:os_version": ["RHEL", "CoreOS"],
-        "another_key": "another_value",
-        "dns-name": "test-instance.ansible.com",
-        "private-dns-name": "test-instance.localhost",
-    }
-
-    inventory._sanitize_hostname = MagicMock()
-    inventory._sanitize_hostname.side_effect = lambda x: x
-
-    m_get_boto_attr_chain.side_effect = lambda pref, instance: instance.get(pref)
-    m_get_tag_hostname.side_effect = lambda pref, instance: instance.get(pref)
-
-    assert expected == inventory._get_all_hostnames(instance, hostnames)
-
-
-def test_inventory_get_all_hostnames_failure(inventory):
-    instance = {}
-    hostnames = [{"value": "some_value"}]
-
-    with pytest.raises(AnsibleError) as err:
-        inventory._get_all_hostnames(instance, hostnames)
-    assert "A 'name' key must be defined in a hostnames dictionary." in str(err.value)
-
-
 @patch("ansible_collections.amazon.aws.plugins.inventory.aws_ec2._get_ssm_information")
 def test_inventory__add_ssm_information(m_get_ssm_information, inventory):
     instances = [
@@ -831,119 +719,3 @@ def test__get_instance_route53_hostnames(inventory, instance, hostnames):
     assert sorted(hostnames) == sorted(result)
 
 
-@patch("ansible_collections.amazon.aws.plugins.inventory.aws_ec2.describe_availability_zones")
-def test_describe_azs_by_region(m_describe_availability_zones, inventory):
-    mock_client = MagicMock()
-    inventory.client = MagicMock(return_value=mock_client)
-
-    m_describe_availability_zones.return_value = [
-        {"ZoneName": "us-east-1a", "ZoneId": "use1-az1", "State": "available"},
-        {"ZoneName": "us-east-1b", "ZoneId": "use1-az2", "State": "available"},
-        {"ZoneName": "us-east-1c", "ZoneId": "use1-az3", "State": "available"},
-    ]
-
-    result = inventory._describe_azs_by_region("us-east-1")
-
-    assert result == {
-        "us-east-1a": {"ZoneName": "us-east-1a", "ZoneId": "use1-az1", "State": "available"},
-        "us-east-1b": {"ZoneName": "us-east-1b", "ZoneId": "use1-az2", "State": "available"},
-        "us-east-1c": {"ZoneName": "us-east-1c", "ZoneId": "use1-az3", "State": "available"},
-    }
-    inventory.client.assert_called_once_with("ec2", region="us-east-1")
-    m_describe_availability_zones.assert_called_once_with(mock_client)
-
-
-@patch("ansible_collections.amazon.aws.plugins.inventory.aws_ec2.describe_availability_zones")
-def test_describe_azs_by_region_caching(m_describe_availability_zones, inventory):
-    mock_client = MagicMock()
-    inventory.client = MagicMock(return_value=mock_client)
-
-    m_describe_availability_zones.return_value = [
-        {"ZoneName": "us-east-1a", "ZoneId": "use1-az1", "State": "available"},
-    ]
-
-    # First call
-    result1 = inventory._describe_azs_by_region("us-east-1")
-    # Second call should use cache
-    result2 = inventory._describe_azs_by_region("us-east-1")
-
-    assert result1 == result2
-    # describe_availability_zones should only be called once due to caching
-    m_describe_availability_zones.assert_called_once()
-    inventory.client.assert_called_once()
-
-
-@patch("ansible_collections.amazon.aws.plugins.inventory.aws_ec2.describe_availability_zones")
-def test_get_availability_zone_ids(m_describe_availability_zones, inventory):
-    inventory.all_clients = MagicMock()
-    inventory.all_clients.return_value = [
-        (MagicMock(), "us-east-1"),
-        (MagicMock(), "us-west-2"),
-    ]
-
-    m_describe_availability_zones.side_effect = [
-        [
-            {"ZoneName": "us-east-1a", "ZoneId": "use1-az1"},
-            {"ZoneName": "us-east-1b", "ZoneId": "use1-az2"},
-        ],
-        [
-            {"ZoneName": "us-west-2a", "ZoneId": "usw2-az1"},
-            {"ZoneName": "us-west-2b", "ZoneId": "usw2-az2"},
-        ],
-    ]
-
-    inventory.client = MagicMock(side_effect=lambda service, region: MagicMock())
-
-    result = inventory._get_availability_zone_ids(strict_permissions=False)
-
-    assert result == {
-        "us-east-1a": "use1-az1",
-        "us-east-1b": "use1-az2",
-        "us-west-2a": "usw2-az1",
-        "us-west-2b": "usw2-az2",
-    }
-    inventory.all_clients.assert_called_once_with("ec2")
-
-
-@patch("ansible_collections.amazon.aws.plugins.inventory.aws_ec2.describe_availability_zones")
-def test_get_availability_zone_ids_with_permissions_error(m_describe_availability_zones, inventory):
-    inventory.all_clients = MagicMock()
-    inventory.all_clients.return_value = [
-        (MagicMock(), "us-east-1"),
-        (MagicMock(), "us-west-2"),
-    ]
-
-    # First region succeeds, second fails with permissions error
-    def describe_side_effect(client):
-        if not hasattr(describe_side_effect, "call_count"):
-            describe_side_effect.call_count = 0
-        describe_side_effect.call_count += 1
-
-        if describe_side_effect.call_count == 1:
-            return [{"ZoneName": "us-east-1a", "ZoneId": "use1-az1"}]
-        else:
-            raise AnsibleInventoryPermissionsError(
-                message="Failed to describe availability zones (permission denied)",
-                exception=botocore.exceptions.ClientError(
-                    {
-                        "Error": {"Code": "UnauthorizedOperation", "Message": "Not authorized"},
-                        "ResponseMetadata": {"HTTPStatusCode": 403},
-                    },
-                    "describe_availability_zones",
-                ),
-            )
-
-    m_describe_availability_zones.side_effect = describe_side_effect
-    inventory.client = MagicMock(side_effect=lambda service, region: MagicMock())
-
-    # With strict_permissions=False, should continue on permissions errors
-    result = inventory._get_availability_zone_ids(strict_permissions=False)
-    assert result == {"us-east-1a": "use1-az1"}
-
-    # With strict_permissions=True, should raise
-    # Clear cache and reset side effect
-    inventory._availability_zone_cache = {}
-    m_describe_availability_zones.side_effect = describe_side_effect
-    describe_side_effect.call_count = 0
-    with pytest.raises(AnsibleInventoryPermissionsError):
-        inventory._get_availability_zone_ids(strict_permissions=True)
